align readme to template
