+++ research/mobilenet/requirements.txt
@@ -1,20 +1,9 @@
-google-api-python-client
-google-cloud-bigquery
-kaggle
-mlperf_compliance==0.0.10
-numpy
-oauth2client
-pandas
-scipy
-tensorflow>=2.1.0
+tensorflow>=2.2.0
 tensorflow-hub
 tensorflow-model-optimization
 tensorflow-datasets>=3.1.0
 tensorflow-addons
 dataclasses
 gin-config
-sentencepiece
 matplotlib
-opencv-python-headless
 pyyaml
-Pillow
