+++ research/mobilenet/tf1_loader/v3_loader.py
@@ -13,29 +13,31 @@
 # limitations under the License.
 # =============================================================================
 
-"""Convert TF v1 MobilenetV2 to TF v2 Keras.
+"""Convert TF v1 MobilenetV3 to TF v2 Keras.
 
 The checkpoint can be found here.
 https://github.com/tensorflow/models/tree/master/research/slim/nets/mobilenet
 
 """
 
-from typing import Text
+from typing import Text, Union
 
-from research.mobilenet import mobilenet_v2_model
+from research.mobilenet import mobilenet_v3
 
 from research.mobilenet.configs import archs
 from research.mobilenet.tf1_loader import utils
 
-MobileNetV2Config = archs.MobileNetV2Config
+MobileNetV3Config = Union[archs.MobileNetV3SmallConfig,
+                          archs.MobileNetV3LargeConfig]
 
 
-def mobinetv2_tf1_tf2_name_convert(tf2_layer_name: Text) -> Text:
+def mobinetv3_tf1_tf2_name_convert(tf2_layer_name: Text) -> Text:
   """Convert TF2 layer name to TF1 layer name. Examples:
   Conv2d_0 -> Conv
   Conv2d_0/batch_norm -> Conv/BatchNorm
-  Conv2d_18 -> Conv_1
-  Conv2d_18/batch_norm -> Conv_1/BatchNorm
+  Conv2d_16 -> Conv_1
+  Conv2d_16/batch_norm -> Conv_1/BatchNorm
+  top/Conv2d_1x1 ->  Conv_2
 
   expanded_conv_1/project -> expanded_conv/project
   expanded_conv_1/depthwise -> expanded_conv/depthwise
@@ -43,6 +45,8 @@ def mobinetv2_tf1_tf2_name_convert(tf2_layer_name: Text) -> Text:
   expanded_conv_2/expand/batch_norm -> expanded_conv_1/expand/BatchNorm
   expanded_conv_2/project -> expanded_conv_1/project
   expanded_conv_2/depthwise -> expanded_conv_1/depthwise
+  expanded_conv_5/squeeze_excite/squeeze -> expanded_conv_4/squeeze_excite/Conv
+  expanded_conv_5/squeeze_excite/excite -> expanded_conv_4/squeeze_excite/Conv_1
 
   top/Conv2d_1x1_output -> Logits/Conv2d_1c_1x1
 
@@ -55,6 +59,8 @@ def mobinetv2_tf1_tf2_name_convert(tf2_layer_name: Text) -> Text:
 
   if 'top/Conv2d_1x1_output' in tf2_layer_name:
     tf1_layer_name = 'Logits/Conv2d_1c_1x1'
+  elif 'top/Conv2d_1x1' in tf2_layer_name:
+    tf1_layer_name = 'Conv_2'
   else:
     if 'batch_norm' in tf2_layer_name:
       tf2_layer_name = tf2_layer_name.replace('batch_norm', 'BatchNorm')
@@ -81,6 +87,13 @@ def mobinetv2_tf1_tf2_name_convert(tf2_layer_name: Text) -> Text:
       raise ValueError('The layer number and type combination is not '
                        'supported: {}, {}'.format(layer_type, str(layer_num)))
 
+    # process squeeze_excite layer
+    if 'squeeze_excite' in ''.join(reminder):
+      if reminder[-1] == 'squeeze':
+        reminder[-1] = 'Conv'
+      elif reminder[-1] == 'excite':
+        reminder[-1] = 'Conv_1'
+
     if target_num:
       tf1_layer_name = '/'.join(['_'.join([layer_type, target_num])] + reminder)
     else:
@@ -89,9 +102,9 @@ def mobinetv2_tf1_tf2_name_convert(tf2_layer_name: Text) -> Text:
   return tf1_layer_name
 
 
-def load_mobilenet_v2(
+def load_mobilenet_v3(
     checkpoint_path: Text,
-    config: MobileNetV2Config = MobileNetV2Config()
+    config: MobileNetV3Config
 ):
   """Load the weights stored in a TF1 checkpoint to TF2 Keras model.
 
@@ -104,18 +117,23 @@ def load_mobilenet_v2(
   """
 
   include_filters = ['ExponentialMovingAverage']
-  exclue_filters = ['RMSProp', 'global_step', 'loss']
+  exclue_filters = ['RMSProp', 'global_step', 'loss', 'Momentum']
   order_keras_weights = utils.generate_layer_weights_map(
     checkpoint_path=checkpoint_path,
     include_filters=include_filters,
     exclude_filters=exclue_filters,
     use_mv_average=True)
 
-  mobilenet_model = mobilenet_v2_model.mobilenet_v2(config=config)
+  if isinstance(config, archs.MobileNetV3LargeConfig):
+    mobilenet_model = mobilenet_v3.mobilenet_v3_large(config=config)
+  elif isinstance(config, archs.MobileNetV3SmallConfig):
+    mobilenet_model = mobilenet_v3.mobilenet_v3_small(config=config)
+  else:
+    raise ValueError('Only support MobileNetV3S and MobileNetV3L')
 
   utils.load_tf2_keras_model_weights(
     keras_model=mobilenet_model,
     weights_map=order_keras_weights,
-    name_map_fn=mobinetv2_tf1_tf2_name_convert)
+    name_map_fn=mobinetv3_tf1_tf2_name_convert)
 
   return mobilenet_model
